+++ lib/flserial.dart
@@ -29,6 +29,7 @@ final DynamicLibrary _dylib = () {
 /// The bindings to the native functions in [_dylib].
 final FlserialBindings _bindings = FlserialBindings(_dylib);
 
+/// Native pointer to string conversion
 String nativeInt8ToString(Pointer<Int8> pointer, {bool allowMalformed = true}) {
   var ptrName = pointer.cast<Utf8>();
   final ptrNameCodeUnits = pointer.cast<Uint8>();
@@ -36,6 +37,7 @@ String nativeInt8ToString(Pointer<Int8> pointer, {bool allowMalformed = true}) {
   return utf8.decode(list, allowMalformed: allowMalformed);
 }
 
+/// String conversion to native pointer
 Pointer<Char> stringToNativeInt8(String str, {Allocator allocator = calloc}) {
   final units = utf8.encode(str);
   final result = allocator<Uint8>(units.length + 1);
@@ -45,6 +47,7 @@ Pointer<Char> stringToNativeInt8(String str, {Allocator allocator = calloc}) {
   return result.cast<Char>();
 }
 
+/// List conversion to native pointer
 Pointer<Char> int8ListToPointerInt8(Uint8List units,
     {Allocator allocator = calloc}) {
   final pointer = allocator<Uint8>(units.length + 1); //blob
@@ -54,6 +57,7 @@ Pointer<Char> int8ListToPointerInt8(Uint8List units,
   return pointer.cast<Char>();
 }
 
+/// Native pointer conversion to list
 Int8List nativeInt8ToInt8List(Pointer<Int8> pointer) {
   var ptrName = pointer.cast<Utf8>();
   final ptrNameCodeUnits = pointer.cast<Int8>();
@@ -61,19 +65,22 @@ Int8List nativeInt8ToInt8List(Pointer<Int8> pointer) {
   return list;
 }
 
+/// Open serial port status
 enum FLOpenStatus {
   open,
   closed,
   error,
 }
 
+/// Class for incoming data handling
 class FlSerialEventArgs {
   FlSerialEventArgs(this.serial, this.len, this.cts, this.dsr);
   final FlSerial serial;
   final int len;
-  final bool cts, dsr;  
+  final bool cts, dsr;
 }
 
+/// Main serial wrapper class over ffi
 class FlSerial {
   int flh = -1;
   List<int> readBuff = List.empty(growable: true);
@@ -81,16 +88,20 @@ class FlSerial {
   bool prevDSR = false;
   var onSerialData = StreamController<FlSerialEventArgs>();
 
+  /// Init should be called at program start, after FlSerial creation
+  /// Function is used to make array of internal port structs for 16 parallel processing ports
   int init() {
     return _bindings.fl_init(16);
   }
 
+  /// Setting callback to native code. Used for incoming data signal
   void setCallback(int flh, {required DartflcallbackFunction callback}) {
     final nativeCallable =
         NativeCallable<flcallbackFunction>.listener(callback);
     _bindings.fl_set_callback(flh, nativeCallable.nativeFunction);
   }
 
+  /// Listing serial ports. Platform depend
   static List<String> listPorts() {
     List<String> list = List<String>.empty(growable: true);
 
@@ -109,33 +120,37 @@ class FlSerial {
     return list;
   }
 
+  /// Internal function for serial port handler checking
   int _checkFLH(int handler) {
     if (handler >= 0 && handler < MAX_PORT_COUNT) {
       return FlError.FL_ERROR_OK;
     }
-    throw FlserialException(FlError.FL_ERROR_HANDLER);
+    throw FlserialException(FlError.FL_ERROR_HANDLER, msg: "$handler");
   }
 
+  /// Function for openning serial port
   FLOpenStatus openPort(String portname, int baudrate) {
-    
     prevCTS = false;
     prevDSR = false;
     flh = _bindings.fl_open(flh, stringToNativeInt8(portname), baudrate);
     int error = _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_LAST_ERROR, -1);
-    if (error >  0) {
-      
-      String msg = "Port open error: " + portname ;
-      switch(error) {
+    int flerror = FlError.FL_ERROR_UNKNOW;
+    if (error > 0) {
+      /// error mapping
+      String msg = "Port open error: " + portname;
+      switch (error) {
         case 1:
-        msg+= " port not exist";
-        break;
+          flerror = FlError.FL_ERROR_PORT_NOT_EXIST;
+          msg += " port not exist";
+          break;
         case 2:
-        msg+= " port allready open";
+          flerror = FlError.FL_ERROR_PORT_ALLREADY_OPEN;
+          msg += " port allready open";
       }
 
       _bindings.fl_close(flh);
       flh = -1;
-      throw Exception(msg);
+      throw FlserialException(flerror, msg: msg);
     }
 
     onSerialData = StreamController<FlSerialEventArgs>();
@@ -145,15 +160,16 @@ class FlSerial {
       callback: (cflh, attrs) async {
         int read = await _readProcess();
         if (read > 0) {
-          onSerialData.add(FlSerialEventArgs(this,
-              readBuff.length, prevCTS, prevDSR));
+          onSerialData
+              .add(FlSerialEventArgs(this, readBuff.length, prevCTS, prevDSR));
         }
       },
     );
-    
+
     return isOpen();
   }
 
+  /// Function check is serial port opened
   FLOpenStatus isOpen() {
     if (flh < 0) {
       return FLOpenStatus.closed;
@@ -170,6 +186,7 @@ class FlSerial {
     return FLOpenStatus.open;
   }
 
+  /// Get last error message. Return empty string if no error
   String getLastError() {
     _checkFLH(flh);
 
@@ -193,9 +210,9 @@ class FlSerial {
     return strres;
   }
 
+  /// Internal function for async serial port reading.
   Future<int> _readProcess() async {
     _checkFLH(flh);
-    // await _readLineStatus();
 
     Uint8List list = await _readList(1024);
 
@@ -203,7 +220,8 @@ class FlSerial {
     return Future<int>.value(readBuff.length);
   }
 
-  /*Future<int> _readLineStatus() async {
+/*
+  Future<int> _readLineStatus() async {
 
     _checkFLH(flh);
 
@@ -220,6 +238,7 @@ class FlSerial {
     return 0;
   }*/
 
+  /// Internal function to read serial data from lower layer with desired len. If no data is present, empty list is returned
   Future<Uint8List> _readList(int len) async {
     _checkFLH(flh);
 
@@ -236,10 +255,12 @@ class FlSerial {
     return list;
   }
 
+  /// Read full read bufer from serial port
   Uint8List readList() {
     return readListLen(readBuff.length);
   }
 
+  /// Read desired len from serial port buffer
   Uint8List readListLen(int len) {
     _checkFLH(flh);
 
@@ -254,6 +275,7 @@ class FlSerial {
     return old;
   }
 
+  /// Read single byte
   int read() {
     _checkFLH(flh);
 
@@ -263,12 +285,14 @@ class FlSerial {
     return -1;
   }
 
-  int write(int len, Uint8List data) {
+  /// Write list data to serial port
+  int write(Uint8List data) {
     _checkFLH(flh);
 
-    return _bindings.fl_write(flh, len, int8ListToPointerInt8(data));
+    return _bindings.fl_write(flh, data.length, int8ListToPointerInt8(data));
   }
 
+  /// Close serial port and free resources
   int closePort() {
     _checkFLH(flh);
     onSerialData.close();
@@ -279,116 +303,139 @@ class FlSerial {
     return flh;
   }
 
+  /// Uniwersal function to control serial port
   int ctrl(int cmd, int value) {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, cmd, value);
   }
 
+  /// Function set RTS line in serial port
   int setRTS(bool value) {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_RTS, value ? 1 : 0);
   }
 
+  /// Function set CTS line in serial port
   bool getCTS() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_GET_CTS, 0) > 0 ? true : false;
   }
 
+  /// Get status of DTR line
   int setDTR(bool value) {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_DTR, value ? 1 : 0);
   }
 
+  /// Get status of DSR line
   bool getDSR() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_GET_DSR, 0) > 0 ? true : false;
   }
 
+  /// Set byte size to 5 bits
   int setByteSize5() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_BYTESIZE_5, 0);
   }
 
+  /// Set byte size to 6 bits
   int setByteSize6() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_BYTESIZE_6, 0);
   }
 
+  /// Set byte size to 7 bits
   int setByteSize7() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_BYTESIZE_7, 0);
   }
 
+  /// Set byte size to 8 bits
   int setByteSize8() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_BYTESIZE_8, 0);
   }
 
-  int setByteParityNone() {
+  /// Set bit parity to none
+  int setBitParityNone() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_PARITY_NONE, 0);
   }
 
-  int setByteParityOdd() {
+  /// Set bit parity to odd
+  int setBitParityOdd() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_PARITY_ODD, 0);
   }
 
-  int setByteParityEven() {
+  // Set bit parity to even
+  int setBitParityEven() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_PARITY_EVEN, 0);
   }
 
-  int setByteParityMark() {
+  // Set bit parity to mark
+  int setBitParityMark() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_PARITY_MARK, 0);
   }
 
-  int setByteParitySpace() {
+  // Set bit parity to space
+  int setBitParitySpace() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_PARITY_SPACE, 0);
   }
 
+  /// Set stop bits to one
   int setStopBits1() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_STOPBITS_ONE, 0);
   }
 
+  /// Set stop bits to one and half
   int setStopBits1_5() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(
         flh, FlCtrl.FL_CTRL_SET_STOPBITS_ONE_POINT_FIVE, 0);
   }
 
+  /// Set stop bits to two
   int setStopBits2() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_STOPBITS_TWO, 0);
   }
 
+  /// Disable flow control
   int setFlowControlNone() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_FLOWCONTROL_NONE, 0);
   }
 
+  /// Set hardware flow control
   int setFlowControlHardware() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_FLOWCONTROL_HARDWARE, 0);
   }
 
+  /// Set xon/xoff flow control (software)
   int setFlowControlSoftware() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_SET_FLOWCONTROL_SOFTWARE, 0);
   }
 
+  /// Waiting for status line change
   int setWaitStatusChange() {
     _checkFLH(flh);
     return _bindings.fl_ctrl(flh, FlCtrl.FL_CTRL_GET_STATUS_CHANGE, 0);
   }
 
+  /// Free resources. Serial port should not be used after calling this function
   int free() {
     return _bindings.fl_free();
   }
 
+  /// Ultility function for get time in milliseconds
   int getTickCount() {
     return DateTime.now().millisecond;
   }