+++ lib/main.dart
@@ -1,2 +1,2 @@
-export 'flserial.dart';
 export 'flserial_exception.dart';
+export 'flserial.dart';